Improve lesson 06 challenges docs and add links to result score

+++ codility-java/src/main/java/codility/lesson06/NumberOfDiscIntersections.java
@@ -39,36 +39,36 @@ import java.util.Arrays;
  * expected worst-case time complexity is O(N*log(N));<br>
  * expected worst-case space complexity is O(N) (not counting the storage
  * required for input arguments).
- * 
- * @author Foroni Marco
+ * @see <a href="https://app.codility.com/programmers/lessons/6-sorting/number_of_disc_intersections/">
+ *     app.codility.com/programmers/lessons/6-sorting/number_of_disc_intersections/</a>
  *
+ * @author Marco Foroni
  */
 final class NumberOfDiscIntersections {
 
 	private static final int MAX_INTERSECTIONS = 10_000_000;
 
 	/**
-	 * Time complexity is O(2*N + N*log(N)) = O(N*log(N))<br>
-	 * Space complexity is O(N) but can be reduced to O(1)
-	 * <p>
-	 * Idea: consider the points that intersect the X axis, if len(A) = N there are
-	 * 2 * N points. Separate the disc starting points from the ending ones using
-	 * two arrays named respectively s and e. Sort these arrays. Then we can define:
-	 * <p>
+	 * Idea: consider the points that intersect the X axis, if len(A) = N there are 2 * N points. Separate the disc
+	 * starting points from the ending ones using two arrays named respectively s and e. Sort these arrays. Then we
+	 * can define:
 	 * i, j, k <- 0
-	 * <p>
+	 * <pre>
 	 * p[i] = min(s[j], e[k])
-	 * <p>
-	 * oc[i] = open circles at position p[i] on the X axis = 
+	 *
+	 * oc[i] = open circles at position p[i] on the X axis =
 	 * oc[i - 1] + 1 if p[i] in s
 	 * oc[i - 1] - 1 if p[i] in e
-	 * <p>
+	 *
 	 * inters[i] = total intersections till position p[i] on the X axis = 
 	 * inters[i - 1] + oc[i - 1] if p[i] in s
 	 * inters[i - 1] otherwise, if p[i] in e
-	 * 
-	 * @param A
-	 * @return
+	 * </pre>
+	 * Time complexity is O(2*N + N*log(N)) = O(N*log(N))<br>
+	 * Space complexity is O(N) but can be reduced to O(1)
+	 *
+	 * @see <a href="https://app.codility.com/demo/results/trainingAGZ4NX-HC7/">
+	 *     app.codility.com/demo/results/trainingAGZ4NX-HC7</a>
 	 */
 	public int solution(int[] A) {
 		if (A.length == 0) {
@@ -82,7 +82,7 @@ final class NumberOfDiscIntersections {
 		}
 		Arrays.sort(start);
 		Arrays.sort(end);
-		// System.out.println(String.format("start=%s, end=%s", MoreInts.toString(start), MoreInts.toString(end)));
+		//System.out.printf("start=%s, end=%s%n", Arrays.toString(start), Arrays.toString(end));
 		final int[] oc = new int[2 * A.length]; // oc[i] = open circles at position i on x axis
 		final int[] inters = new int[2 * A.length]; // inters[i] = number of intersections at position i on X axis
 		assert (start[0] <= end[0]);
@@ -103,7 +103,7 @@ final class NumberOfDiscIntersections {
 				return -1;
 			}
 			i++;
-			// System.out.println(String.format("inters=%s, o.c.=%s", MoreInts.toString(inters), MoreInts.toString(oc)));
+			//System.out.printf("inters=%s, o.c.=%s%n", Arrays.toString(inters), Arrays.toString(oc));
 		}
 		return inters[2 * A.length - 1];
 	}
@@ -111,9 +111,6 @@ final class NumberOfDiscIntersections {
 	/**
 	 * Time complexity is O(N^2)<br>
 	 * Space complexity is O(1)
-	 * 
-	 * @param A
-	 * @return
 	 */
 	public static int exhaustiveSearch(final int[] A) {
 		int counter = 0;
@@ -131,20 +128,20 @@ final class NumberOfDiscIntersections {
 	}
 
 	/**
-	 * FIXME: https://app.codility.com/demo/results/trainingX2GPUS-NHD/ 93% not 100%
-	 * FIXME: https://app.codility.com/demo/results/trainingUDA9BS-SJE/ 93% not 100%
+	 * Time complexity is O(N*logN)<br>
+	 * Space complexity is O(N)
 	 *
-	 * @param A
-	 * @return
+	 * @see <a href="https://app.codility.com/demo/results/trainingNXCFPG-RUB/">
+	 *     app.codility.com/demo/results/trainingNXCFPG-RUB</a>
 	 */
 	public static int elegantSolution(int[] A) {
 		final Point[] points = new Point[A.length * 2];
 		for (int i = 0; i < A.length; i++) {
-			points[2 * i] = new Point(i - A[i], true);
-			points[2 * i + 1] = new Point(i + A[i], false);
+			points[2 * i] = new Point(i - (long)A[i], true);
+			points[2 * i + 1] = new Point(i + (long)A[i], false);
 		}
 		Arrays.sort(points);
-		// System.out.println(Arrays.toString(points));
+		//System.out.println(Arrays.toString(points));
 		int discIntersections = 0;
 		int openDiscs = 0;
 		for (Point p : points) {
@@ -157,7 +154,7 @@ final class NumberOfDiscIntersections {
 			if (discIntersections > 10_000_000) {
 				return -1;
 			}
-			// System.out.println(String.format("Open discs: %d, Intersections: %d.", openDiscs, discIntersections));
+			//System.out.printf("Open discs: %d, Intersections: %d.%n", openDiscs, discIntersections);
 		}
 		return discIntersections;
 	}
